requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
sqlalchemy
psycopg2-binary
alembic 